requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==21.1
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.6
wheel==0.37.0
environ~=1.0
gunicorn==20.0.4
psycopg2-binary



future~=0.18.2
olefile~=0.46
pip~=20.0.2
keyring~=18.0.1
numpy~=1.21.2
setuptools~=45.2.0
colorama~=0.4.3
PyYAML~=5.3.1
Click~=7.0
httplib2~=0.14.0
pandas~=1.3.3
python-dateutil~=2.7.3
SQLAlchemy~=1.3.24
simplejson~=3.16.0
blinker~=1.4
chardet~=3.0.4
bcrypt~=3.1.7
django-email-users~=1.1.0